% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswissmaps-package.R
\docType{package}
\name{ggswissmaps-package}
\alias{ggswissmaps}
\alias{ggswissmaps-package}
\title{ggswissmaps: Offers Various Swiss Maps as Data Frames and 'ggplot2' Objects}
\description{
Offers various swiss maps as data frames and 'ggplot2' objects and gives the possibility to add layers of data on the maps. Data are publicly available from the swiss federal statistical office. In addition to the \code{maps2} object (a list of 8 swiss maps, at various levels), there are the data frames with the boundaries used to produce these maps (\code{shp_df}, a list with 8 data frames).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gibonet/ggswissmaps}
}

}
\author{
\strong{Maintainer}: Sandro Burri \email{gibo.gaf@gmail.com}

Authors:
\itemize{
  \item Eric Stephani \email{eric.stephani@bluewin.ch}
}

Other contributors:
\itemize{
  \item Lisa Bottinelli \email{lisa.bottinelli@gmail.com} [contributor]
}

}
\keyword{internal}
