% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\docType{package}
\name{ggswissmaps}
\alias{ggswissmaps}
\title{Offers various swiss maps as data frames and 'ggplot2' objects.}
\description{
Offers various swiss maps as data frames and 'ggplot2' objects 
and gives the possibility to add layers of data on the maps. 
Data are publicly available from the swiss federal statistical office. 
In addition to the \code{maps2} object (a list of 8 swiss maps, at 
various levels), there are the data frames with the boundaries used to
produce these maps (\code{shp_df}, a list with 8 data frames).
}
\details{
See the package vignette for some examples 
(\code{vignette("ggswissmaps_intro", package = "ggswissmaps")}).
}
