% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\name{maps}
\alias{maps}
\title{A list with 8 maps of Switzerland's territory at various levels.}
\format{A list with 16 ggplot2 maps with swiss territory}
\source{
\url{http://www.bfs.admin.ch/bfs/portal/fr/index/dienstleistungen/geostat/datenbeschreibung.html}
}
\description{
Every element of the list is a ggplot2 object which is
already ready to plot.
\itemize{
  \item distretti. Swiss districts
  \item comuni. Swiss communes (2014)
  \item boh. Don't know
  \item boh2. Don't know 2
  \item cantoni. Swiss cantons
  \item ch. Swiss national boundaries
  \item gr. Swiss big regions (NUTS3)
}
}
\examples{
data(maps)
maps[[1]]
names(maps)
maps[["cantoni"]]
str(maps[["cantoni"]])
str(maps[["cantoni"]]$data)
}

