% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\name{shp_df}
\alias{shp_df}
\title{A list with 8 data frames of swiss territory boundaries, at various levels.}
\format{
A list with 8 data frames with swiss territory boundaries (at various levels).
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/limites-administratives/limites-communales-generalisees.html}. Download date: 2015-08-18
}
\description{
Every element of the list is a data frame, which can be used with 'ggplot2'.
}
\details{
Columns are not all the same across data frames, but usually they have the
following in common:
\itemize{
  \item long. Longitude coordinate (x)
  \item lat. Latitude coordinate (y)
  \item group. A factor to be used to plot the polygons correctly (with ggplot2)
}
}
\examples{
data(shp_df)
class(shp_df)
length(shp_df)
names(shp_df)
str(shp_df[["g1k15"]])
}
