% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\name{shp_sf}
\alias{shp_sf}
\title{A list with 8 'sf' objects of swiss territory boundaries, at various levels.}
\format{
A list with 8 'sf' data frames with swiss territory boundaries (at various levels).
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/services/geostat/geodonnees-statistique-federale/limites-administratives/limites-communales-generalisees.html}. Download date: 2015-08-18
}
\description{
Every element of the list is a 'sf' data frame, which can be used with 'ggplot2'.
}
\details{
Every 'sf' data frame has a column "geometry", which is a list-column containing
the boundaries of the polygons.

See the vignette (\code{vignette("ggswissmaps_with_sf", package = "ggswissmaps")}) 
for some examples.
}
\examples{
data(shp_sf)
class(shp_sf)
length(shp_sf)
names(shp_sf)
str(shp_sf[["g1k15"]])
}
