% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\docType{data}
\name{theme_white}
\alias{theme_white}
\title{theme_white is a ggplot2 theme object and can be added to a ggplot2 object
to eliminate axes, ticks and put white background}
\format{\preformatted{List of 8
 $ axis.title      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.text       : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.ticks      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.line       : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.background: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.border    : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.margin    : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
theme_white
}
\description{
theme_white is a ggplot2 theme object and can be added to a ggplot2 object
to eliminate axes, ticks and put white background
}
\keyword{datasets}

