% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\docType{package}
\name{ggswissmaps}
\alias{ggswissmaps}
\alias{ggswissmaps-package}
\title{Offers various swiss maps as ggplot2 objects.}
\description{
Offers various swiss maps as ggplot2 objects and gives the possibility
to add layers of data on the maps. Data are publicly available from the
swiss federal statistical office.
}
\details{
There are only two objects in package \pkg{ggswissmaps}:
\code{theme_white} (for white background of ggplot2 objects) and \code{maps} (a list of 8 swiss maps).
}

