% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggswissmaps_data.R
\docType{data}
\name{theme_white}
\alias{theme_white}
\title{theme_white is a ggplot2 theme object and can be added to a ggplot2 object  
to eliminate axes, ticks and put white background}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 8.}
\usage{
theme_white
}
\description{
theme_white is a ggplot2 theme object and can be added to a ggplot2 object  
to eliminate axes, ticks and put white background
}
\keyword{datasets}

